requirements.txt
django==4.2.17
djangorestframework==3.16.0
drf-spectacular==0.28.0
flake8==7.2.0
mysqlclient==2.2.7
python-dotenv==1.1.0
